services/permit-ingester/requirements.txt
fastapi>=0.111
uvicorn>=0.30
httpx>=0.27
pyyaml>=6.0
tenacity>=8.4
